+++ src/main/java/org/opensearch/knn/jni/NmslibService.java
@@ -11,7 +11,7 @@
 
 package org.opensearch.knn.jni;
 
-import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.BuiltinKNNEngine;
 import org.opensearch.knn.index.query.KNNQueryResult;
 import org.opensearch.knn.index.store.IndexInputWithBuffer;
 import org.opensearch.knn.index.store.IndexOutputWithBuffer;
@@ -35,7 +35,7 @@ class NmslibService {
     static {
         KNNLibraryLoader.loadNmslibLibrary();
         initLibrary();
-        KNNEngine.NMSLIB.setInitialized(true);
+        BuiltinKNNEngine.NMSLIB.setInitialized(true);
     }
 
     /**