+++ src/test/java/org/opensearch/knn/index/query/KNNWeightTests.java
@@ -40,6 +40,7 @@ import org.opensearch.knn.index.codec.KNN990Codec.QuantizationConfigKNNCollector
 import org.opensearch.knn.index.codec.KNNCodecVersion;
 import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 import org.opensearch.knn.index.codec.util.KNNVectorAsCollectionOfFloatsSerializer;
+import org.opensearch.knn.index.engine.BuiltinKNNEngine;
 import org.opensearch.knn.index.engine.KNNEngine;
 import org.opensearch.knn.index.engine.MethodComponentContext;
 import org.opensearch.knn.index.quantizationservice.QuantizationService;
@@ -102,7 +103,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
             testQueryScore(
                 space::scoreTranslation,
                 SEGMENT_FILES_NMSLIB,
-                Map.of(SPACE_TYPE, space.getValue(), KNN_ENGINE, KNNEngine.NMSLIB.getName())
+                Map.of(SPACE_TYPE, space.getValue(), KNN_ENGINE, BuiltinKNNEngine.NMSLIB.getName())
             );
         }
     }
@@ -116,7 +117,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
                 SPACE_TYPE,
                 SpaceType.L2.getValue(),
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 PARAMETERS,
                 String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
             )
@@ -129,7 +130,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
                 SPACE_TYPE,
                 SpaceType.INNER_PRODUCT.getValue(),
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 PARAMETERS,
                 String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
             )
@@ -143,7 +144,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
                 SPACE_TYPE,
                 SpaceType.INNER_PRODUCT.getValue(),
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 PARAMETERS,
                 String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
             )
@@ -162,7 +163,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
 
         ModelDao modelDao = mock(ModelDao.class);
         ModelMetadata modelMetadata = mock(ModelMetadata.class);
-        when(modelMetadata.getKnnEngine()).thenReturn(KNNEngine.FAISS);
+        when(modelMetadata.getKnnEngine()).thenReturn(BuiltinKNNEngine.FAISS);
         when(modelMetadata.getSpaceType()).thenReturn(spaceType);
         when(modelMetadata.getState()).thenReturn(ModelState.CREATED);
         when(modelMetadata.getVectorDataType()).thenReturn(VectorDataType.DEFAULT);
@@ -231,7 +232,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
 
         ModelDao modelDao = mock(ModelDao.class);
         ModelMetadata modelMetadata = mock(ModelMetadata.class);
-        when(modelMetadata.getKnnEngine()).thenReturn(KNNEngine.FAISS);
+        when(modelMetadata.getKnnEngine()).thenReturn(BuiltinKNNEngine.FAISS);
         when(modelMetadata.getSpaceType()).thenReturn(spaceType);
 
         KNNWeight.initialize(modelDao);
@@ -408,7 +409,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         final FieldInfo fieldInfo = mock(FieldInfo.class);
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             PARAMETERS,
             String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
         );
@@ -497,7 +498,9 @@ public class KNNWeightTests extends KNNWeightTestCase {
         when(reader.getFieldInfos()).thenReturn(fieldInfos);
         when(fieldInfos.fieldInfo(FIELD_NAME)).thenReturn(fieldInfo);
 
-        when(fieldInfo.attributes()).thenReturn(Map.of(KNN_ENGINE, KNNEngine.FAISS.getName(), SPACE_TYPE, SpaceType.HAMMING.getValue()));
+        when(fieldInfo.attributes()).thenReturn(
+            Map.of(KNN_ENGINE, BuiltinKNNEngine.FAISS.getName(), SPACE_TYPE, SpaceType.HAMMING.getValue())
+        );
 
         FSDirectory directory = mock(FSDirectory.class);
         when(reader.directory()).thenReturn(directory);
@@ -640,7 +643,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         final FieldInfo fieldInfo = mock(FieldInfo.class);
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             SPACE_TYPE,
             isBinary ? SpaceType.HAMMING.getValue() : SpaceType.L2.getValue()
         );
@@ -738,7 +741,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         final FieldInfo fieldInfo = mock(FieldInfo.class);
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             SPACE_TYPE,
             SpaceType.L2.getValue()
         );
@@ -837,7 +840,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
             final KNNWeight knnWeight = new DefaultKNNWeight(query, boost, filterQueryWeight);
             final Map<String, String> attributesMap = ImmutableMap.of(
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 VECTOR_DATA_TYPE_FIELD,
                 isBinary ? VectorDataType.BINARY.getValue() : VectorDataType.FLOAT.getValue(),
                 SPACE_TYPE,
@@ -953,7 +956,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
                 SPACE_TYPE,
                 spaceType.getValue(),
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 PARAMETERS,
                 String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
             )
@@ -1008,7 +1011,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         final KNNWeight knnWeight = new DefaultKNNWeight(query, boost, filterQueryWeight);
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             SPACE_TYPE,
             SpaceType.L2.name(),
             PARAMETERS,
@@ -1086,7 +1089,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         final KNNWeight knnWeight = new DefaultKNNWeight(query, boost, filterQueryWeight);
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             SPACE_TYPE,
             SpaceType.L2.name(),
             PARAMETERS,
@@ -1174,7 +1177,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
             // Mock field Info
             final Map<String, String> attributesMap = ImmutableMap.of(
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 SPACE_TYPE,
                 SpaceType.L2.name(),
                 PARAMETERS,
@@ -1200,7 +1203,16 @@ public class KNNWeightTests extends KNNWeightTestCase {
             // Then
             // Verifies ANN index was called
             jniServiceMockedStatic.verify(
-                () -> JNIService.queryIndex(anyLong(), any(float[].class), eq(K), anyMap(), eq(KNNEngine.FAISS), any(), anyInt(), any())
+                () -> JNIService.queryIndex(
+                    anyLong(),
+                    any(float[].class),
+                    eq(K),
+                    anyMap(),
+                    eq(BuiltinKNNEngine.FAISS),
+                    any(),
+                    anyInt(),
+                    any()
+                )
             );
             jniServiceMockedStatic.verifyNoMoreInteractions();
         }
@@ -1251,7 +1263,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
             final KNNWeight knnWeight = new DefaultKNNWeight(query, boost, filterQueryWeight);
             final Map<String, String> attributesMap = ImmutableMap.of(
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 SPACE_TYPE,
                 SpaceType.HAMMING.name(),
                 PARAMETERS,
@@ -1485,7 +1497,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         final FieldInfo fieldInfo = mock(FieldInfo.class);
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             SPACE_TYPE,
             SpaceType.L2.getValue()
         );
@@ -1584,7 +1596,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
                 SPACE_TYPE,
                 SpaceType.L2.getValue(),
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 PARAMETERS,
                 String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
             )
@@ -1653,7 +1665,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
         // Prepare fieldInfo
         final Map<String, String> attributesMap = ImmutableMap.of(
             KNN_ENGINE,
-            KNNEngine.FAISS.getName(),
+            BuiltinKNNEngine.FAISS.getName(),
             SPACE_TYPE,
             SpaceType.L2.name(),
             PARAMETERS,
@@ -1723,8 +1735,8 @@ public class KNNWeightTests extends KNNWeightTestCase {
         when(fieldInfos.fieldInfo(any())).thenReturn(fieldInfo);
         when(fieldInfo.attributes()).thenReturn(fileAttributes);
 
-        String engineName = fieldInfo.attributes().getOrDefault(KNN_ENGINE, KNNEngine.NMSLIB.getName());
-        KNNEngine knnEngine = KNNEngine.getEngine(engineName);
+        String engineName = fieldInfo.attributes().getOrDefault(KNN_ENGINE, BuiltinKNNEngine.NMSLIB.getName());
+        KNNEngine knnEngine = BuiltinKNNEngine.getEngine(engineName);
         List<String> engineFiles = KNNCodecUtil.getEngineFiles(knnEngine.getExtension(), query.getField(), reader.getSegmentInfo().info);
         String expectIndexPath = String.format("%s_%s_%s%s%s", SEGMENT_NAME, 2011, FIELD_NAME, knnEngine.getExtension(), "c");
         assertEquals(engineFiles.get(0), expectIndexPath);
@@ -1775,7 +1787,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
             final FieldInfo fieldInfo = mock(FieldInfo.class);
             final Map<String, String> attributesMap = ImmutableMap.of(
                 KNN_ENGINE,
-                KNNEngine.FAISS.getName(),
+                BuiltinKNNEngine.FAISS.getName(),
                 PARAMETERS,
                 String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
             );
@@ -1858,7 +1870,7 @@ public class KNNWeightTests extends KNNWeightTestCase {
                 final FieldInfo fieldInfo = mock(FieldInfo.class);
                 final Map<String, String> attributesMap = ImmutableMap.of(
                     KNN_ENGINE,
-                    KNNEngine.FAISS.getName(),
+                    BuiltinKNNEngine.FAISS.getName(),
                     PARAMETERS,
                     String.format(Locale.ROOT, "{\"%s\":\"%s\"}", INDEX_DESCRIPTION_PARAMETER, "HNSW32")
                 );